% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-set.R
\docType{data}
\name{common_names}
\alias{common_names}
\title{common_names}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{name}{First name}
  \item{gender}{Gender (Male or Female)}
  \item{per_1000}{Frequency per 1,000 Americans}
}
}
\source{
\url{https://fivethirtyeight.com/features/whats-the-most-common-name-in-america/}
}
\usage{
common_names
}
\description{
The data-set that contains common names in the US during 2013
}
\keyword{datasets}
